fix(proxy): satisfy stream settlement CI gates
